+++ src/main/kotlin/com/gacagebot/domain/repositories/TenorApiRepository.kt
@@ -1,0 +1,10 @@
+package com.gacagebot.domain.repositories
+
+import com.gacagebot.domain.models.TenorMediaFormats
+import io.reactivex.rxjava3.core.Single
+
+interface TenorApiRepository {
+
+    fun fetchTenorGif(query: String, limit: Int, random: Boolean): Single<TenorMediaFormats>
+
+}